Fix module structure
Moved functions from onedim_utils to onedim_core that are used by
MatrixProducState methods.

Also some other minor fixes.
