+++ tncontract/onedim/__init__.py
@@ -5,9 +5,4 @@ onedimensional
 Subpackage for one dimensional tensor networks
 """
 from tncontract.onedim.onedim_core import *
-import tncontract.onedim.mps_methods
-
-
-#from tncontract.onedim.one_dimension import(init_mps_random, 
-#        tensor_to_mps, tensor_to_mpo, expvals_mps, onebody_sum_mpo)
-
+import tncontract.onedim.onedim_utils
